+++ CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this extension will be documented in this file.
   output channel, or by running the `Confluent: Show Output Channel` command.
 - Right-click actions on Confluent/Kafka resources to easily copy their IDs, names (if applicable),
   and URLs (if applicable) to the clipboard.
+- Text search fields in Message Viewer that works along with partition filter providing intersection
+  of filtered results. This first iteration of text search is case sensitive and substring-only
+  (i.e. no "exact" value matching).
+- Button to open currently consumed messages as a JSON file.
 
 ### Changed